+++ src/algorithms/f.rs
@@ -7,39 +7,56 @@ enum FState {
     Integrate(f32),
 }
 
-pub struct F<const SAMPLES_300: usize, const SAMPLES_50: usize> {
+pub struct F<FMW, FB> {
+    fs: SamplingFrequency,
+
     /// F should be initialized at the same time as M is, skip earlier samples
     state: FState,
+
     /// 300ms of the individual max samples of the 50ms buffer.
     /// We store the maxima to save some memory and computation.
-    f_max_window: SlidingWindow<f32, [f32; SAMPLES_300]>,
+    f_max_window: SlidingWindow<f32, FMW>,
+
     /// 50ms window of the signal
-    f_buffer: SlidingWindow<f32, [f32; SAMPLES_50]>,
+    f_buffer: SlidingWindow<f32, FB>,
 }
 
-impl<const SAMPLES_300: usize, const SAMPLES_50: usize> F<SAMPLES_300, SAMPLES_50> {
-    pub fn new(fs: SamplingFrequency) -> Self {
+impl<FMW, FB> F<FMW, FB>
+where
+    FMW: AsRef<[f32]> + AsMut<[f32]>,
+    FB: AsRef<[f32]> + AsMut<[f32]>,
+{
+    pub fn new(
+        fs: SamplingFrequency,
+        f_max_window: SlidingWindow<f32, FMW>,
+        f_buffer: SlidingWindow<f32, FB>,
+    ) -> Self {
         // sanity check buffer sizes
         debug_assert_eq!(
-            SAMPLES_300,
+            f_max_window.capacity(),
             fs.ms_to_samples(300.0),
-            "Incorrect type parameters, must be <{}, {}>",
-            fs.ms_to_samples(300.0),
-            fs.ms_to_samples(50.0)
+            "Incorrect FMW type parameter. Buffer must be {} samples long.",
+            fs.ms_to_samples(300.0)
         );
         debug_assert_eq!(
-            SAMPLES_50,
+            f_buffer.capacity(),
             fs.ms_to_samples(50.0),
-            "Incorrect type parameters, must be <{}, {}>",
-            fs.ms_to_samples(300.0),
+            "Incorrect FB type parameter. Buffer must be {} samples long.",
             fs.ms_to_samples(50.0)
         );
 
         Self {
+            fs,
             state: FState::Ignore(fs.s_to_samples(2.65)),
-            f_max_window: SlidingWindow::default(),
-            f_buffer: SlidingWindow::default(),
+            f_max_window,
+            f_buffer,
         }
+    }
+
+    pub fn clear(&mut self) {
+        self.state = FState::Ignore(self.fs.s_to_samples(2.65));
+        self.f_max_window.clear();
+        self.f_buffer.clear();
     }
 
     fn update_f_buffers(&mut self, sample: f32) -> (Option<f32>, f32) {
@@ -62,14 +79,14 @@ impl<const SAMPLES_300: usize, const SAMPLES_50: usize> F<SAMPLES_300, SAMPLES_5
 
     pub fn update(&mut self, sample: f32) {
         self.state = match self.state {
-            FState::Ignore(1) => FState::Init(SAMPLES_300 - 1, 0.0),
+            FState::Ignore(1) => FState::Init(self.f_max_window.capacity() - 1, 0.0),
             FState::Ignore(n) => FState::Ignore(n - 1),
             FState::Init(n, favg) => {
                 let favg = favg + sample;
                 self.update_f_buffers(sample);
 
                 if n == 0 {
-                    FState::Integrate(favg.max(0.0) / (SAMPLES_300 as f32))
+                    FState::Integrate(favg.max(0.0) / (self.f_max_window.capacity() as f32))
                 } else {
                     FState::Init(n - 1, favg)
                 }